Improved

+++ Transformation/transformation/src/main/scala/getisOrd/SoH.scala
@@ -9,15 +9,15 @@ import geotrellis.raster.Tile
 class SoH {
     def getSoHDowAndUp(parent : Tile, child : Tile, clusterOfParent : Int, clusterOfChild : Int): (Double, Double) ={
       val childParent = (new ClusterRelations()).getNumberChildrenAndParentsWhichIntersect(parent,child)
-      val down = childParent._2/clusterOfParent
-      val up = childParent._1/clusterOfChild
+      val down = childParent._2.toDouble/clusterOfParent.toDouble
+      val up = childParent._1.toDouble/clusterOfChild.toDouble
       (down, up)
     }
 
   def getSoHDowAndUp(tuple : ((Tile,Int),(Tile,Int))): (Double, Double) ={
     val childParent = (new ClusterRelations()).getNumberChildrenAndParentsWhichIntersect(tuple._1._1,tuple._2._1)
-    val down = childParent._2/tuple._1._2
-    val up = childParent._1/tuple._2._2
+    val down = childParent._2.toDouble/tuple._1._2.toDouble
+    val up = childParent._1.toDouble/tuple._2._2.toDouble
     (down, up)
   }
 }
